Build and install scheduler target as a module library
Installing scheduler as an interface obfuscates some useful information
about the target from the build system. An example of which is that
the `TARGET_FILE` generator expression does is not supported by
interface libraries.

VCPKG expects an import library to be installed alongside shared
libraries, which is the reason why scheduler was previously exported
as an interface library.

Since Scheduler is not intended to be linked against, but may be
dynamically loaded, we can build it as a module library in order to
address both of these issues.